Fixed USART interrupt bug for vTaskInput
Incorrectly assumed interrupt was not firing but in fact it was consistently firing not allowing any other process to run. This was because the USART3->RDR register was not being read and thus not clearing the interrupt causing the ISR to continuously trigger. Another issue with NVIC priorities was identified which also prevented the interrupt from behaving correctly. On Cortex-M chips the priority of an interrupt by default is 0, greater than the allowed limit of as defined in FreeRTOSConfig.h - manually increase the NVIC priority to a value greater than or equal to configLIBRARY_MAX_SYSCALL_INTERRUPT_PRIORITY.