Updated blink trigger from vTaskDelay to TIM2 Interrupt
Delay caused switching from a blinking state to a disabled or enabled state is still present without vTaskDelay. Could possibly be a combination of waiting for bits but also have a wait (semaphore or delay) nested in the same function. Incorporation of TIM2 is not wasted as this can lead to custom blink intervals once UART can receive more than one character